+++ snippy/content/collection.py
@@ -320,15 +320,25 @@ class Collection(object):  # pylint: disable=too-many-public-methods
 
         return text
 
-    def dump_term(self, use_ansi, debug_logs):
-        """Convert collection for terminal."""
+    def dump_term(self, templates, template_format, use_ansi, debug_logs):
+        """Convert collection for terminal.
+
+        Args:
+           templates (dict): Dictionary that contains content templates.
+           template_format (str): Define the output format
+           use_ansi (bool): Define if ANSI characters are used.
+           debug_logs (bool): Define if debut information is included.
+        """
 
         if not self:
             Cause.push(Cause.HTTP_NOT_FOUND, 'cannot find content with given search criteria')
 
         text = Const.EMPTY
-        for i, resource in enumerate(self.resources(), start=1):
-            text = text + resource.dump_term(index=i, use_ansi=use_ansi, debug_logs=debug_logs)
+        if template_format == Const.CONTENT_FORMAT_MKDN:
+            text = self.dump_mkdn(templates)
+        else:
+            for i, resource in enumerate(self.resources(), start=1):
+                text = text + resource.dump_term(index=i, use_ansi=use_ansi, debug_logs=debug_logs)
 
         # Set one empty line at the end of string for beautified output.
         if self: